+++ src/commands/deployWorkspaceProject/getDefaultValues/getDefaultContextValues.ts
@@ -8,8 +8,6 @@ import { parseAzureResourceId } from "@microsoft/vscode-azext-azureutils";
 import { nonNullValue, type ISubscriptionActionContext } from "@microsoft/vscode-azext-utils";
 import { ImageSource } from "../../../constants";
 import { ext } from "../../../extensionVariables";
-import { type SetTelemetryProps } from "../../../telemetry/SetTelemetryProps";
-import { type DeployWorkspaceProjectTelemetryProps as TelemetryProps } from "../../../telemetry/commandTelemetryProps";
 import { ContainerAppItem } from "../../../tree/ContainerAppItem";
 import { ManagedEnvironmentItem } from "../../../tree/ManagedEnvironmentItem";
 import { localize } from "../../../utils/localize";
@@ -22,7 +20,7 @@ import { getDefaultContainerAppsResources } from "./getDefaultContainerAppsResou
 import { getWorkspaceProjectPaths } from "./getWorkspaceProjectPaths";
 
 export async function getDefaultContextValues(
-    context: ISubscriptionActionContext & SetTelemetryProps<TelemetryProps>,
+    context: ISubscriptionActionContext & Partial<DeployWorkspaceProjectContext>,
     item?: ContainerAppItem | ManagedEnvironmentItem
 ): Promise<Partial<DeployWorkspaceProjectContext>> {
     const { rootFolder, dockerfilePath } = await getWorkspaceProjectPaths(context);
